- (Diep) - Changed bestiary tab to morphology tab and renamed a couple files.
